refactor(ui): modularize interface components
- Split UI components into separate modules:
  - sidebar.py for navigation and thread management
  - css.py for styling and theme definitions
  - chat_interface.py for chat interaction components
- Improve code organization and maintainability
- Reduce main file complexity (chatbot.py)
